+++ addOns/authhelper/CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 - Do not fail the authentication on diagnostic errors.
+- Do not configure poll authentication verification without logged in indicator.
 
 ## [0.27.0] - 2025-07-03
 ### Added